+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Documentation in English and German ([#17](https://github.com/scm-manager/scm-editor-plugin/pull/17))
+- Add extension point to show hints for files below the editor ([#20](https://github.com/scm-manager/scm-editor-plugin/pull/20))
 
 ## 2.0.0 - 2020-06-04
 ### Changed